+++ lib/PipedrivePUT.rb
@@ -1,5 +1,5 @@
-require "PipedrivePUT/version"
-require "PipedrivePUT/organization"
+require 'PipedrivePUT/version'
+require 'PipedrivePUT/organization'
 require 'PipedrivePUT/search'
 require 'PipedrivePUT/users'
 require 'PipedrivePUT/deal'
@@ -11,12 +11,18 @@ require 'PipedrivePUT/activity'
 require 'PipedrivePUT/activity-type'
 require 'PipedrivePUT/currencies'
 require 'PipedrivePUT/deal_fields'
-
+require 'PipedrivePUT/filters'
+require 'PipedrivePUT/files'
 
 require 'json'
 require 'open-uri'
 require 'rest-client'
+require 'money'
+require 'money/bank/google_currency'
 
+
+#this is for exchange calculations to make currency format nicely.
+I18n.enforce_available_locales = false
 
 module PipedrivePUT