+++ lib/PipedrivePUT/currencies.rb
@@ -2,16 +2,26 @@ module PipedrivePUT
   class Currencies
     include PipedrivePUT
 
-#---------------------------------------------------------get all currencies----------------------------------------------------------------------------------------------------------------------
+
+    def self.getExchangeRate(currency_name, options = {})
+      get_curr = PipedrivePUT::Currencies.getCurr(currency_name)
+      get_code = get_curr['data'][0]['code']
+      Money::Bank::GoogleCurrency.ttl_in_seconds = 86400
+      Money.default_bank = Money::Bank::GoogleCurrency.new
+      get_amount = options[:amount].to_i
+      set_new_exchange = options[:ex_code].parameterize.to_sym
+      set_amount_and_code_to_exchange = Money.new(get_amount, get_code)
+      exchange_rate = set_amount_and_code_to_exchange
+      formated_rate = exchange_rate.exchange_to(set_new_exchange).format
+    end
+    
     def self.getAllCurr
       @base = 'https://api.pipedrive.com/v1/currencies?api_token=' + @@key.to_s
       @content = open(@base.to_s).read
       @parsed = JSON.parse(@content)
       return @parsed
     end 
-#---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
-#----------------------------------------- search for currencies -----------------------------------------------------------------------------------------------------------------------------------
     def self.getCurr(currency_name)
       curr_string = currency_name.to_s 
       curr_string_count = curr_string.split.size
@@ -61,8 +71,7 @@ module PipedrivePUT
         @parsed = JSON.parse(@content)
         return @parsed
       end 
-    end
-#----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
+    end 
   end
 end